Prepare to replace _dataset_fit
The way the experiment fetches models: Try to load
them, if it's not on disk - train, save and return
the trained model.

If a specific epoch is requested but not found this
is a more serious error - the calling experiment
obviously expects the model to already be trained
if a specific epoch is needed. This is reflected
in the _get_model_at_epoch version which raises an
exception.

Signed-off-by: dorimedini <dorimedini@gmail.com>